database 根据调价获取卡牌加个添加筛选条件。jhs cli 更新商品逻辑优化，根据价格区间直接从数据库直接获取待更新商品。

+++ internal/database/models/card_price.go
@@ -30,7 +30,7 @@ type CardPrice struct {
 
 // 卡牌价格的查询条件
 type CardPriceQuery struct {
-	CardVersionID  int64    `json:"card_version_id"`
+	CardVersionID  int      `json:"card_version_id"`
 	SetsPrefix     []string `json:"set_prefix"`
 	Color          []string `json:"color"`
 	Keyword        string   `json:"keyword"`
